Refactor into event and speaker packages
