--- sampleData/content/events/2018-canterlot/speakers/apple-jack.md
@@ -1,15 +1,0 @@
-+++
-Title = "Apple Jack"
-type = "speaker"
-
-twitter = "applejack"
-
-
-
-
-
-+++
-Apple Jack is awesome.
-
-so is *markdown*
-
