+++ sampleData/content/events/2018-canterlot/speakers/rainbow-dash.md
@@ -1,0 +1,12 @@
++++
+Title = "Rainbow Dash"
+type = "speaker"
+
+
+
+
+
+
+
++++
+
