+++ sampleData/data/events/2018-canterlot.yml
@@ -1,7 +1,7 @@
 name: "2018-canterlot" # The name of the event. Four digit year with the city name in lower-case, with no spaces.
 year: "2018" # The year of the event. Make sure it is in quotes.
 city: "Canterlot" # The displayed city name of the event. Capitalize it.
-event_twitter: "canterlot" # Change this to the twitter handle for your event such as devopsdayschi or devopsdaysmsp
+event_twitter: "" # Change this to the twitter handle for your event such as devopsdayschi or devopsdaysmsp
 description: "Devopsdays is coming to Canterlot!" # Edit this to suit your preferences
 ga_tracking_id: "" # If you have your own Google Analytics tracking ID, enter it here. Example: "UA-74738648-1"
